refactor: reorganize authentication and chat components, update hooks and state management
- Moved authentication hooks to use-auth.ts for better organization.
- Refactored chat components to use new desktop and mobile sidebar components.
- Implemented a search modal for user searching functionality.
- Updated user chat input and space components to utilize new hooks.
- Created modal and sheet stores for managing UI state.
- Improved code readability and structure across various components.
