+++ frontend/src/hooks/use-auth.ts
