+++ frontend/src/hooks/use-debounce.ts
@@ -1,0 +1,15 @@
+import { useEffect, useState } from "react";
+
+export function useDebounce(value: string, delay = 500) {
+  const [debounceValue, setDebounceValue] = useState(value);
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setDebounceValue(value);
+    }, delay);
+
+    return () => clearTimeout(timer);
+  }, [value, delay]);
+
+  return debounceValue;
+}
