+++ frontend/src/shared/chats/user-chat-space.tsx
@@ -1,22 +1,13 @@
-import {
-  IconMenu2,
-  IconSettingsFilled,
-  IconUserCircle,
-  IconX,
-} from "@tabler/icons-react";
+import { IconMenu2, IconSettingsFilled, IconUserCircle, IconX } from "@tabler/icons-react";
 import { useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
 import { Spinner } from "../../components/ui/spinner";
-import { useFetchMessages, useReceiverDetails } from "../../hooks/user-hook";
+import { useFetchMessages, useReceiverDetails } from "../../hooks/use-user";
 import { useAuthState } from "../../store/auth-store";
+import { useSheetStore } from "../../store/modal-store";
 import UserChatInput from "./user-chat-input";
 
-interface UserChatNavbarProps {
-  isOpen: boolean;
-  setIsOpen: (open: boolean) => void;
-}
-
-const UserChatSpace = ({ isOpen, setIsOpen }: UserChatNavbarProps) => {
+const UserChatSpace = () => {
   const {
     getReceiverDetails,
     data: dataReceiver,
@@ -31,6 +22,7 @@ const UserChatSpace = ({ isOpen, setIsOpen }: UserChatNavbarProps) => {
   } = useFetchMessages();
 
   const user = useAuthState((s) => s.user);
+  const { openSheet, isSheetOpen } = useSheetStore();
 
   const [searchParams] = useSearchParams();
   const selectedChat = searchParams.get("chat");
@@ -45,9 +37,7 @@ const UserChatSpace = ({ isOpen, setIsOpen }: UserChatNavbarProps) => {
   if (!selectedChat) {
     return (
       <div className="bg-black h-screen flex items-center justify-center">
-        <p className="text-gray-500 text-xl font-medium">
-          Select a chat to start message.
-        </p>
+        <p className="text-gray-500 text-xl font-medium">Select a chat to start message.</p>
       </div>
     );
   }
@@ -104,12 +94,8 @@ const UserChatSpace = ({ isOpen, setIsOpen }: UserChatNavbarProps) => {
         <div className="flex-row flex items-center gap-2">
           <IconSettingsFilled className="text-gray-400" size="28" />
           {/* Mobile Menu Toggle Button */}
-          <button
-            type="button"
-            onClick={() => setIsOpen(!isOpen)}
-            className="flex md:hidden"
-          >
-            {isOpen ? (
+          <button type="button" onClick={openSheet} className="flex md:hidden">
+            {isSheetOpen ? (
               <IconX className="text-gray-400" size={28} />
             ) : (
               <IconMenu2 className="text-gray-400" size={28} />
@@ -120,34 +106,38 @@ const UserChatSpace = ({ isOpen, setIsOpen }: UserChatNavbarProps) => {
 
       {/* Chat Messages */}
       <div className="flex-1 overflow-y-auto px-8 py-6 space-y-4">
-        {dataMessages?.messages.map((message) => {
-          const isMine = user?.id === message.sender_id;
+        {dataMessages?.conversationId ? (
+          dataMessages?.messages.map((message) => {
+            const isMine = user?.id === message.sender_id;
 
-          return (
-            <div
-              key={message.id}
-              className={`flex gap-3 ${
-                isMine ? "flex-row-reverse" : "justify-start"
-              }`}
-            >
-              <div className="w-10 h-10 rounded-full overflow-hidden">
-                <img
-                  alt="Tailwind CSS chat bubble component"
-                  src="https://img.daisyui.com/images/profile/demo/kenobee@192.webp"
-                />
-              </div>
+            return (
               <div
-                className={`chat-bubble max-w-xs lg:max-w-md px-4 py-2 rounded-lg ${
-                  isMine
-                    ? "bg-primary-100/50 text-white rounded-br-none"
-                    : "bg-gray-800 text-gray-100 rounded-bl-none"
-                }`}
+                key={message.id}
+                className={`flex gap-3 ${isMine ? "flex-row-reverse" : "justify-start"}`}
               >
-                {message.content}
+                <div className="w-10 h-10 rounded-full overflow-hidden">
+                  <img
+                    alt="Tailwind CSS chat bubble component"
+                    src="https://img.daisyui.com/images/profile/demo/kenobee@192.webp"
+                  />
+                </div>
+                <div
+                  className={`chat-bubble max-w-xs lg:max-w-md px-4 py-2 rounded-lg ${
+                    isMine
+                      ? "bg-primary-100/50 text-white rounded-br-none"
+                      : "bg-gray-800 text-gray-100 rounded-bl-none"
+                  }`}
+                >
+                  {message.content}
+                </div>
               </div>
-            </div>
-          );
-        })}
+            );
+          })
+        ) : (
+          <div>
+            <p className="text-white">no chats found</p>
+          </div>
+        )}
 
         {loadingMessages && (
           <div className="flex items-center mt-10 justify-center bg-black h-screen">
